Fix Int>Float type in response data

+++ app/src/main/java/no/uio/ifi/in2000/adrianch/adrianch/skumring/model/sunrise/SunriseResponseData.kt
@@ -1,6 +1,6 @@
 package no.uio.ifi.in2000.adrianch.adrianch.skumring.model.sunrise
 
-data class Geometry(val coordinates: List<Int>,
+data class Geometry(val coordinates: List<Float>,
                     val type: String = "")
 
 data class Properties(val solarnoon: Solarnoon,
